// apps/backend/src/application/controllers/trainings/training-entity.interface.ts
import { UUID } from '../../../types/uuid.type'

export interface TrainingEntityInterface {
  id: UUID;
  name: string;
  backgroundImgFileName: string;
  trainingLevel: string;
  trainingType: string;
  trainingDuration: string,
  price: number;
  calories: number;
  description: string;
  sex: string;
  videoDemoFileName: string;
  rating: number;
  votesNumber: number;
  trainingCreatorId: UUID;
  isSpecial: boolean;
}